Give GuiWindow a dummy "engine" collaborator
Obviously the window isn't where the HTTP requests or layout calculations
happen, so after a bit of back-and-forth, the object that coordinates all that
is called the engine :/

We can launch and quit the Gosu application right now, but I haven't really
found a reliable way to automate user interaction with it that'll run in
different desktop environments.

It seems like the best way to keep this project moving forward would be to
instantiate either a GuiWindow or a stateful backing object of GuiWindow's and
spy on its collaborators to verify it's doing what we expect. Not end-to-end,
but as close to without a serious amount of work unrelated to the app itself