Run app under test in a thread
This seems to work. I might've cargo-culted this into existence a little.

It seems like we *need* to do it this way, because to perform any sort of
drawing with Gosu, the app needs to be running and triggering the
GuiWindow#draw callback. Any less than relying on that callback would be less
"integrated" and has already proven to happily hide a whole bunch of bugs