NodeFactory creates either a Node or a TextNode
This is a messier change than I'd originally imagined. Some of our incoming
(from the real world!) text content contains line breaks and it seems our
drawing library chokes on line breaks.

Text nodes are a bit tricky. We know there isn't a one-to-one mapping between
DOM nodes and render tree nodes, especially when it comes to text, but this
(the initial transition from DOM to render tree) might be too early a stage to
break them up into separate lines.

Even if it wasn't too early, this probably isn't the place for it. So, note to
self: move this when the shape of the layout pipeline becomes clearer