Move responsibility for tree traversal to visitor
This is a difficult one. I initially thought, wishfully, that a visitor object
doesn't have to be tightly coupled to their tree traversal strategy. Turns out
in a lot of cases they probably do.

That isn't to say this couldn't be done more elegantly or be extracted into
something more reusable later on, mind