Move #content to TextNode's public interface
Text is a funny, messy thing in markup. We're soon gonna want a parent node to
be able to concatenate its children's text content (and for that process to be
legible :D). E.g. with the following markup

  <p>I read <a>an excellent blog post</a> recently.</p>

it'd be nice if the P tag could return the whole string from its children and
grandchildren as opposed to having deal with the three text nodes separately