Text search for render tree
There was some debate as to whether we should be working with text nodes
directly or via their parent bounding box. Should text nodes have size and
position of their own? That's how I've modelled it in my puny human brain, so
that's the route we're going down for now. This may well turn out to be messy
/ overly complicated / etc