Introduce more useful rendering tests
We've removed the GUI-launching from our features and now we're trying to test
useful things about the state of the laid-out render tree "as we believe it
will be drawn".

This is a decent start in that direction and is thanks to @bestie. It's the
first in a possible group of tests, this one assuring us of the approximate
position of our main heading on the page