Always create drawing visitor within #draw thread
There is information that will change during the application's lifetime and the
drawing objects will need the most up-to-date version of. The kind of
information I'm thinking of right now is "viewport" size.

Other than the obvious choice of updating this information from the #draw
method because that's when it's needed, I think Gosu also has a problem
retrieving things like its width and height outside of the draw thread