Use partial doubles to prevent more mishaps
This is an attempt at supplementing the existing, not particularly end-to-end,
specs. I've already managed to merge a branch that broke the "build" because
the specs aren't exercising the integration between the application code, the
Gosu adapters and Gosu's own objects.

Making use of RSpec's partial double verification is all new to me, so maybe
this isn't the optimal choice, but I thought it was worth a go. My go-to
approach would usually be to define a double in its own file and run a shared
example to verify the interface of both the double and the mocked object*.

* Specs verifying the real object like this might cross over into "slow
  integration test" territory, but in theory there would be so few of those to
  run that it wouldn't slow the suite down by much