Rename method for clarity
This method isn't drawing anything! A visitor is given a root node, it asks
that root node to make use of a type-specific callback (type being Node or
TextNode). That's how a visitor goes about visiting a node. The drawing is
performed in that callback, and when the visitor visits a node's children, it
merely asks them to use their own appropriate callback so that they can also be
drawn