Remove visitor#visit method
It didn't start out this way, but in light of the growing number of visitor
objects we did some simple refactoring and since then this method has been both
identical in all visitors and, specifically, only doing this:

  tree.accept_visit(self)

So, the caller (Engine, GuiWindow, essentially some fairly high level
orchestration sorta object) tells the visitor to visit the tree, like

  visitor.visit(tree)

and the visitor in turn does little more than delegating the start of the visit
back to the tree.

It's hard to say which makes more sense. In the Gang of Four description of the
visitor pattern, there's actually a visit-orchestrating "object structure", so
either way our current options are both lacking that