Relax definition of traverser in shared example
In reality it turns out a visitor can traverse the render tree in a depth-first
fashion, supplying any argument it wishes to the nodes' #accept_visit method.

It'd be great if I could figure out custom RSpec matchers and ensure that the
argument is, say, `a_visitor` (i.e. conforms to the "a visitor" shared
example). I seem to have a lot of trouble navigating RSpec's docs, though