Introduce IntrinsicWidthSetter
For now we're using this to set the widths of lines of text.

I'm just gonna go with this and see if we can implement text wrapping later...
it might be interesting, or impossible.

In a more advanced layout engine, "intrinsic" might only apply to inflexible
objects like images, I'm not really clear on that at this point