Rework/rename children-positioning
The children-ness of this activity is less important once you're inside the
object itself. At that point, we're just dealing with a collection of nodes.

On top of that, the ideas of "position" and "positioning" are overloaded in the
project and are no longer helpful. This object works with a group of nodes, two
consecutive siblings at a time and negotiates between them to arrange them
either inline or block-level