Use LayoutPipeline utility for drawing visitors
We've had this in place for the LAYOUT_VISITORS for a long time, but there
hasn't been any need to do this for drawing too until now.

I think it might be time to break up the drawing visitors. It _might_ be less
efficient at first but it should increase readability quite a bit