Introduce FailingFetcher that fails everytime
Our OfflineHtmlFetcher's been handy for working with web pages offline
(obvs) and also for tests, but now that we have image fetching things
get a little more complicated. I think it'll make sense to use different
fetchers for stubbing the initial page download and the later requests
for assets. This is my first OK idea to that effect