Move dimensions calculation into the visitor
That is, move it from ImageStore into ImageFetcher.

This kinda thing's the job of the visitor in all other cases so far. "This
kinda thing" I'm imagining to include any use of GuiWindow's facilities that
I can list right now off the top of my head :D

And this step feels like it really removes some complexity