Prevent access to local files outside of "origin"
The idea is that this would prevent unintended arbitrary file system
traversal. Maybe that just doesn't matter if we are in a genuine,
guaranteed offline state (i.e. unable to leak arbitrary local data to
any remote location).

So on those grounds I'm not confident this move is actually a positive
one. At the very least I was interested to see how it could be
introduced.

When in offline mode, the browser is either asked to render a blob of
HTML or the contents of a named HTML file. In the former case, the
"origin" is the working directory of the project itself. In the latter
case, the origin is the directory of the named file