Allow text width stub to be configured in specs
The offline engine's had this configurability for a while now but the
regular engine hasn't.

Most of the existing feature specs use the offline engine but I'd like
to try moving those feature specs over to the regular engine.

For now I'm just copy-pasting over any differences, but once it's all
working more of this will be revisited