Introduce WindowDouble that's kinda like GuiWindow
So far, there's no interaction with GuiWindow in the specs -- feature
specs have dealt with Engine directly.

1. I'd like it to be easier for the specs to interact with the browser
   as close to the-ways-a-user-would as possible

2. test coverage stops at the boundary of the UI code, which I'm not
   *too* worried about right now, but the more logic we inevitably add
   to the GuiWindow, the less it's strictly UI code and the more
   concerned I'll be

In order to replace Engine in the step definitions, WindowDouble's been
given the bare minimum interface to allow us resize the window
dimensions (kinda... that original words-per-row thing seems like it big
mistake now) and to check that a node is rendered within the viewport.

With any luck, a cleanup of the step definitions will drive the change
to stop returning the render tree as-is from WindowDouble#go. It feels
like a messy coupling. Though, in a way, it's a reasonably well defined
interface, I think there's a difference between the internal render tree
and existing real world, standardised browser interfaces, e.g. a DOM