Reinstate Text's internal `box`, remove `position`
The problem with any node's width and height (and as a result right and
bottom) being dynamically calculated from its children, is the
difference in behaviour after the render tree's been transformed by the
AbsolutePositionAdderUpper.

That doesn't impact Text nodes right now, since TextRows' absolute
positions are resolved on the fly by the drawing visitor. But I don't
like the assymmetry there, I think this code would be easier to use if
all things that need an absolute position were assigned them in the same
way.

As part of this change, I've moved some code from Arranger to
TextWrapper, to have the related "adjustment" code in one place