Introduce TextSearch object
The immediate purpose of this object is to allow Inspector to provide
more accurate results for the location of text content on the rendered
page.

TextSearch traverses the render tree, extracts the
rendered page's entire text content, and via #find_owners_of will return
all TextRows that contribute one or more characters to the first
occurance of the provided search string.

This is a change from the approach Inspector has taken so far, of
traversing the render tree on-the-fly, using recursion to narrow in on
occurances of the search string -- which ultimately seems to result in
quite clumsy bundles of sibling TextRows whether they're relevant or not