Rephrase and use new Inspector method
Just for the layout scenarios for now.

This reduces the amount of feature support code and I've taken the
opportunity to remove some low value steps / assertions (probably also
assertions that shouldn't've been steps in the first place).

If you've noticed the search string "Youradhere", it illustrates a bug
in text layout that all of these improvements are hopefully, slowly
building towards fixing