Remove #clone_with_children special case method
I'm trying to reduce the number of methods an element decorator is
required to define (that don't relate directly to its M.O.) and
`#clone_with_children` was one of those.

A lot of uses of `#clone_with` do actually change the `children` array
anyway, so the name of this one's bothered me for a bit. Something to
consider when it's time to name it again.

There's also logic equivalent to this for Text (in TextWrapper), so it
could be worth combining those at some point in the future