Point#to_h
I'd like to remove `#position_after` from the list of methods that node
decorators are required to override. That means taking the cloning out
of that method. To achieve that, I want the operation "get the next
available position after node Y and move into that position" to be as
straightforward as possible