Make #position_after #the_position_after
I.e. just return the Point representing the position after the given
node but no longer clone the receiver with the Point's x and y.

The cloning here caused issues for the Ruby delegators which meant that
all node-decorating objects needed to implement #position_after whether
they did anything special with it or not