Script for taking a screenshot of a website
It'll put the screenshot in the tmp/ directory. Or output errors to
stderr and exit with a non-zero status if the browser crashed.

Errors for crashes are actually a mess, especially since the final error
messages I've seen so far relate to X11 issues caused by the early exit
of the browser and therefore Xvfb too.

BTW, `import` is provided by ImageMagick

---

Git wasn't acknowledging the tmp/ and tmp/gitkeep rules added to
gitignore when following the approach taken with assets/, so I've
changed both to match the new style (that works for now)