Stop caching downloaded files
SDL occasionally chokes on a downloaded image.

(FWIW I assume the issue is with the downloading of the image.)

That not only means a failed render but also a crash.

Unfortunately, as long as that page contains a reference to that image,
any future requests for the page will also cause a crash.

Let's revisit caching when there's time to get it right