Specify number of combined words and spaces in spec
We're testing how text wraps at the edge of the screen when it's too
long to fit onto a single row.

The TextWrapper itself doesn't make this distinction between words and
spaces, they're all just chunks of text content to either combine into
a single row or split into multiple rows.

I was so bothered about how much sense this made when you read it out
loud. This change doesn't help in that area, but it might now be a bit
easier to use