Stop resetting TextRow x,y after wrapping text
This fixes a bug.

It's a hangover from mistakenly assuming that inline elements always
consist of a single box (x, y, width, height). This *might* make sense
for block-level layouts, but it turns out that it's no good for inline
flows.

Until this commit, child nodes could only be positioned as far left as
their parent's X position. That's usually fine for text that fits into
the remaining space of the row that it begins on. But not good if it
begins partway across the screen and has to wrap onto multiple rows