Rename test support method
We're trying to separate the concept of the browser app from the GUI
window, and in the process assessing the responsbilities of GuiWindow
and its test environment counterpart, WindowDouble.

Despite its name, this `resize_window` method configures the
TextCalculator (to only allow a certain number of "words" to fit across
the width of the viewport).

Outside of the test environment, TextCalculator is provided by the GUI.
Therefore, the WindowDouble does seem to be the right place for this
method and hopefully the new name makes that clearer