Move drawing pipeline from GuiWindow to Engine
This
* puts more of the browser's behaviour under our control, not the GUI's
* at least partially exercises the drawing pipeline during feature tests
* could allow tests to work with some version of the results of drawing
  to screen, in contrast to the current limitation of working with the
  layout pipeline's render tree (I found this to be a problem when
  investigating how to test-drive scrolling down the page)

---

I think the original motivation for this responsibility to lie with
GuiWindow was that drawing to screen seemed inseparable from the GUI
library and the current state of the window.

In fact, the GUI library is mostly encapsulated in the renderer
utilities that it provides for the drawing pipeline to use. And as
things stand the current state of the window is represented in the
"viewport", which is passed to the Engine everytime we re-render